Refactoring Sidebar
Sidebar divided into components:CreateBoard, NeedHelpBlock, LogOut,Overlay
