+++ src/shared/components/Header/Header.styled.jsx
@@ -14,14 +14,13 @@ export const StyledHeader = styled.header`
   box-sizing: border-box;
   flex-shrink: 0;
   background: ${({ theme }) => theme.colors.backgroundHeader};
-  transition: width 225ms cubic-bezier(0.4, 0, 0.6, 1) 0ms,
+    transition: width 225ms cubic-bezier(0.4, 0, 0.6, 1) 0ms,
     margin 225ms cubic-bezier(0.4, 0, 0.6, 1) 0ms;
 
   @media screen and (min-width: ${tablet}) {
     padding: 18px 32px;
     position: absolute;
-    left: 0;
-  }
+      }
 
   @media screen and (min-width: ${desktop}) {
     padding: 18px 24px;
@@ -55,6 +54,7 @@ export const UserName = styled.p`
   line-height: normal;
   letter-spacing: -0.28px;
   color: ${({ theme }) => theme.colors.textHeader};
+  
 `;
 
 export const AvatarImg = styled.img`
