+++ src/shared/components/Sidebar/Sidebar.jsx
@@ -1,179 +1,43 @@
-import { useDispatch } from "react-redux";
 import PropTypes from "prop-types";
-import {
-  StyledSidebar,
-  AppName,
-  Title,
-  Plus,
-  Wrapper,
-  CreateBoardButton,
-  CreateText,
-  LogoWrapper,
-  BlockWrapper,
-  SupportMessage,
-  MessageWrapper,
-  HelpBtn,
-  HelpIcon,
-  LogOutBtn,
-  LogOutIcon,
-  GreenColor,
-  Overlay,
-} from "./Sidebar.styled";
+import { StyledSidebar, AppName, LogoWrapper } from "./Sidebar.styled";
 import sprite from "../../images/icons.svg";
-import Helper from "../../images/help.png";
+
 import { Logo } from "./Sidebar.styled";
 import { BoardList } from "./BoardList";
-import { signOut } from "../../../redux/auth/operations.js";
-import { useMediaQuery } from "react-responsive";
-import localStorage from "redux-persist/es/storage";
-import { createGlobalStyle } from "styled-components";
+
+import { CreateBoard } from "./CreateBoard/CreateBoard";
+import { NeedHelpBlock } from "./NeedHelpBlock/NeedHelpBlock";
+import { LogOut } from "./LogOut/LogOut";
+
 export const Sidebar = ({
   isOpen,
-  onOpen,
+  onOpenCreateBoard,
   onOpenHelp,
   onOpenEditDashBoard,
-  closeSidebar,
 }) => {
-  const dispatch = useDispatch();
-
-  const handleLogOut = () => {
-    dispatch(signOut());
-  };
-
-  const isDesktopOrLaptop = useMediaQuery({
-    query: "(min-width: 1440px)",
-  });
-
   return (
-    <>
-      {isDesktopOrLaptop && (
-        <StyledSidebar
-          onClick={() => {
-            console.log("Я сайт");
-          }}
-          isOpen={isOpen}
-        >
-          <LogoWrapper>
-            <Logo>
-              <use href={sprite + "#icon-logo"}></use>
-            </Logo>
-            <AppName>Task Pro</AppName>
-          </LogoWrapper>
+    <StyledSidebar isOpen={isOpen}>
+      <LogoWrapper>
+        <Logo>
+          <use href={sprite + "#icon-logo"}></use>
+        </Logo>
+        <AppName>Task Pro</AppName>
+      </LogoWrapper>
 
-          <Title>My boards</Title>
+      <CreateBoard onOpenCreateBoard={onOpenCreateBoard} />
 
-          <Wrapper>
-            <CreateText>
-              Create a<br></br>new board
-            </CreateText>
-            <CreateBoardButton onClick={onOpen}>
-              <Plus>
-                <use href={sprite + "#icon-plus"}></use>
-              </Plus>
-            </CreateBoardButton>
-          </Wrapper>
+      <BoardList onOpenEditDashBoard={onOpenEditDashBoard} />
 
-          <BoardList onOpenEditDashBoard={onOpenEditDashBoard} />
+      <NeedHelpBlock onOpenHelp={onOpenHelp} />
 
-          <BlockWrapper>
-            <img
-              src={Helper}
-              style={{ width: "54px", height: "78px" }}
-              alt="/"
-            />
-
-            <MessageWrapper>
-              <SupportMessage>
-                If you need help with
-                <br /> <GreenColor>TaskPro</GreenColor>, check out our
-                <br /> support resources or <br /> reach out to our <br />{" "}
-                customer support team.
-              </SupportMessage>
-            </MessageWrapper>
-
-            <HelpBtn type="button" onClick={onOpenHelp}>
-              <HelpIcon>
-                <use href={sprite + "#icon-help-circle"} />
-              </HelpIcon>
-              Need help?
-            </HelpBtn>
-          </BlockWrapper>
-
-          <LogOutBtn type="button" onClick={handleLogOut}>
-            <LogOutIcon>
-              <use href={sprite + "#icon-login"} />
-            </LogOutIcon>
-            Log out
-          </LogOutBtn>
-        </StyledSidebar>
-      )}
-      {!isDesktopOrLaptop && (
-        <>
-          <StyledSidebar isOpen={isOpen}>
-            <LogoWrapper>
-              <Logo>
-                <use href={sprite + "#icon-logo"}></use>
-              </Logo>
-              <AppName>Task Pro</AppName>
-            </LogoWrapper>
-
-            <Title>My boards</Title>
-
-            <Wrapper>
-              <CreateText>
-                Create a<br></br>new board
-              </CreateText>
-              <CreateBoardButton onClick={onOpen}>
-                <Plus>
-                  <use href={sprite + "#icon-plus"}></use>
-                </Plus>
-              </CreateBoardButton>
-            </Wrapper>
-
-            <BoardList onOpenEditDashBoard={onOpenEditDashBoard} />
-
-            <BlockWrapper>
-              <img
-                src={Helper}
-                style={{ width: "54px", height: "78px" }}
-                alt="/"
-              />
-
-              <MessageWrapper>
-                <SupportMessage>
-                  If you need help with
-                  <br /> <GreenColor>TaskPro</GreenColor>, check out our
-                  <br /> support resources or <br /> reach out to our <br />{" "}
-                  customer support team.
-                </SupportMessage>
-              </MessageWrapper>
-
-              <HelpBtn type="button" onClick={onOpenHelp}>
-                <HelpIcon>
-                  <use href={sprite + "#icon-help-circle"} />
-                </HelpIcon>
-                Need help?
-              </HelpBtn>
-            </BlockWrapper>
-
-            <LogOutBtn type="button" onClick={handleLogOut}>
-              <LogOutIcon>
-                <use href={sprite + "#icon-login"} />
-              </LogOutIcon>
-              Log out
-            </LogOutBtn>
-          </StyledSidebar>
-          <Overlay onClick={closeSidebar} isOpen={isOpen} />
-        </>
-      )}
-    </>
+      <LogOut />
+    </StyledSidebar>
   );
 };
 
 Sidebar.propTypes = {
   isOpen: PropTypes.bool.isRequired,
-  onOpen: PropTypes.func.isRequired,
+  onOpenCreateBoard: PropTypes.func.isRequired,
   onOpenHelp: PropTypes.func.isRequired,
   onOpenEditDashBoard: PropTypes.func.isRequired,
-  closeSidebar: PropTypes.func.isRequired,
 };
